LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/fetch-event-referrer-policy.https-expected.txt



FAIL Service Worker responds to fetch event with the referrer policy promise_test: Unhandled rejection with value: object "TypeError: Not allowed to request resource"
FAIL Service Worker should respond to fetch with the default referrer policy assert_equals: expected "ReferrerPolicy: strict-origin-when-cross-origin" but got "ReferrerPolicy: no-referrer-when-downgrade"
FAIL Service Worker should respond to fetch with the referrer URL when a member of RequestInit is present - Default Referrer assert_equals: expected "Referrer: https://localhost:9443/service-workers/service-worker/resources/simple.html?referrerPolicy\nReferrerPolicy: strict-origin-when-cross-origin" but got "Referrer: https://localhost:9443/service-workers/service-worker/resources/simple.html?referrerPolicy\nReferrerPolicy: no-referrer-when-downgrade"
